Hotfix: Don't use `Array.find`, support Internet Explorer
This updates the romo javascript to no longer use `Array.find`
because it isn't supported in Internet Explorer. This is part of
an effort to support later versions of Internet Explorer. In this
case this switches the `Romo.children` followed by an `Array.find`
that uses `Romo.is` to instead just pass the selector to the
`Romo.children` method and select the first item from the result.
This has the same result and switches to using `Array.filter` in
the `Romo.children` method instead of `Array.find`.