Hotfix: Have `Romo.width` and `Romo.height` fall back to `Romo.css`
This updates the `Romo.width` and `Romo.height` helpers to fall
back and use `Romo.css` if they can't get a width or height using
the `getBoundingClientRect`. This is to fix issues with trying to
get the height/width of an element that isn't being displayed.
Specifically this fixes pickers and selects appropriately
inheriting their width from their hidden inputs or select
elements.

Note: This does reintroduce an issue with IE since `Romo.css` in
IE doesn't return the expected width or height values when using
the border-box box sizing but it's now limited to only hidden
elements. This will be addressed in a future commit.