Update Modal and DropDown `openPopup`
This updates the `openPopup` method in both the Modal and
Dropdown js. In both cases, the `contenElem` was undefined,
thereby erroring when `outerHTML` was being called on it. This
is erroring live in an app. This is now updated to correctly
find the `contentElem`.

Also, I changed the `outerHTML` call on the `contentElem` to
`innerHtml`. With the "outer" call, nothing was being returned
as the standard mark-up for this situation has a `romo-display-none`
class. Changing to `innerHTML` fixed this issue.

This method, in both the Modal and Dropdown js, were updated in a
previous effort to get the `RomoModal` component to stop using
jquery. For reference see: https://github.com/redding/romo/commit/1e9381e