Update `doStart` and `doStop` to methods in Spinner js
This updates the `doStart` and `doStop` methods in the Spinner js
to include undefined checks before any of their specific logic
is run. This allows for calling start on an already started spinner
and stop on an already stopped spinner. This would error in some
cases when the same spinner elems were triggered to start/stop
multiple times.