(hotfix) bugfix for previous spinner `doStop` change
I accidentally broke the spinner `doStop` method in PR 256.  I
moved the stop logic into a push function so that it would match
the start logic being put in a push function.  The goal was to
let all the accumulated DOM events run before mutation the DOM
on start/stop.

This works for the `doStart` method b/c the spinner return check
and setting the spinner both are not done in the push function.
However, it broke the `doStop` method b/c the spinner return check
was not done in the push function but the spinner clearing was
done in the push function.  You have to change the spinner state
at the same time you do the return check - otherwise subsequent
checks that occur before the push function won't check the correct
state.

To fix this, I just more the return check into the push function.
I have to unset the spinner in the push function so this means I
have to do the return check in the push function.  This retains
the previous behavior of allowing events to propagate before
changing the DOM while keeping accurate spinner state checking.