Hotfix: Update form submit logic to check for buttons and inputs
This update the form submit logic to check for buttons and inputs
as well as elements with the data romo form submit set. This fixes
an error when trying to find a submit elem from a click event.
The logic tries to find the submit elem that triggered the click
by checking if the target is a submit elem or finding the closest
submit elem. The logic was updated to only check for data attr
submit elems in PR 255 but should have been updated to check for
all kinds of submit elems. This updates the form class to know
its submit elem selector and use it in the submit logic to find a
submit elem.