Add `RomoDate.daysDiff` helper
This adds a helper for calculating the number of days between
two dates. The helper converts the dates to their millisecond
time values, subtracts them, and then converts the difference to
a number of days by dividing it by the number of milliseconds in
a day. The helper rounds the difference to get an integer of the
days.