Update form to allow event submitting
This updates the form component to allow configuring it to do an
event submit instead of a browser or ajax submit. The goal is to
provide a way to use romo form but not have it make a request to
the server (either as an standard HTTP request or as an ajax
request).

This adds a data attr for event submitting and checks it whenever
the form is submitted (either via a click or the trigger event).
The event submit simply builds the form values the same as an
ajax submit then triggers an event submit event with the form
values. This will allow listening for the event and then using the
form values as needed. This also proxies the event in the dropdown
form, inline form, and modal form so it can be used with those
components as well.

This also renames the non-browser submit methods to use ajax
instead of non-browser. Since the submit logic is no longer a
binary between browser or non-browser this makes for a better
name (because event submitting is also non-browser submitting).
This also matches the other private method that is already
named `_ajaxSubmit`.

Finally, this also moves the browser submit event before
submitting the form. This is because submitting the form reloads
the page so an event after it can't really be used. This allows
the event to be used before the page reloads.