add base js `.removeChildren` method
This is similar to jQuery's `.empty` method.  We debated about
reproducing that method in Romo and ultimately decided against it
b/c of not being sold on the name and edge-cases with removing
both child nodes and text.

I'm proposing this implementation as it only handles the child
nodes case and is named explicitly as such.  This compliments the
`.updateText` method where you update it to empty string to clear
the text.